+++ src/console/src/Contract/FormatterInterface.php
@@ -1,6 +1,6 @@
-<?php
+<?php declare(strict_types=1);
 
-namespace Swoft\Console\Advanced;
+namespace Swoft\Console\Contract;
 
 /**
  * Interface FormatterInterface