+++ src/framework/src/SwoftApplication.php
@@ -2,6 +2,8 @@
 
 namespace Swoft;
 
+use Swoft\Contract\ApplicationInterface;
+use Swoft\Contract\SwoftInterface;
 use Swoft\Processor\AnnotationProcessor;
 use Swoft\Processor\ApplicationProcessor;
 use Swoft\Processor\BeanProcessor;
@@ -15,6 +17,7 @@ use Swoft\Stdlib\Helper\ComposerHelper;
 
 /**
  * Swoft application
+ * @since 2.0
  */
 class SwoftApplication implements SwoftInterface, ApplicationInterface
 {